+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 ## Unreleased
-- Fix the DaemonSet name on diff which prevented pulumi to replace the resource (https://github.com/pulumi/pulumi-kubernetes/pull/1951)
 (None)
+
+## 3.18.2 (April 6, 2022)
+- Only add keyring default value when verification is turned on (https://github.com/pulumi/pulumi-kubernetes/pull/1961)
+  Regression introduced in 3.18.1
+- Fix the DaemonSet name on diff which prevented pulumi to replace the resource (https://github.com/pulumi/pulumi-kubernetes/pull/1951)
 
 ## 3.18.1 (April 5, 2022)
 - Fix autonaming panic for helm release (https://github.com/pulumi/pulumi-kubernetes/pull/1953)